Verification of print start stop and speed
The multinozzle now can start printing at the correct place and the correct velocity. A list of next steps was made. Many debugging features were added.

+++ src/mjdriver.cpp
@@ -11,6 +11,7 @@
 #include <QBitmap>
 #include <QPainter>
 #include <format>
+#include <QTimer>
 
 #include <opencv2/opencv.hpp>
 #include <nlohmann/json.hpp>
@@ -157,7 +158,7 @@ void Controller::set_head_voltage(HeadIndex idx, double voltage)
     write_line(command.toUtf8());
 }
 
-void Controller::set_absolute_start(double steps)
+void Controller::set_absolute_start(int steps)
 {
     if (steps <= 0) return;
     QString command = QString(", %1").arg(steps);
